Put the class into the global namespace. Renamed the file and the class itself.
